Renamed added pawn class
